Fixed type size warnings.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@9577 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/CtrlHomePanel.cpp
@@ -716,7 +716,7 @@ void CHomeLibraryBox::Update()
 		}
 	}
 	
-	size_t nHashing = LibraryBuilder.GetRemaining();
+	DWORD nHashing = (DWORD)LibraryBuilder.GetRemaining();
 	
 	if ( nHashing > 0 )
 	{
